Refactor onboarding views in new architecture
